feat: move the same filed into same file
