--- src/components/SearchBar/fields/AdvancedDateMobileField.tsx
@@ -1,52 +1,0 @@
-import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger,
-} from '@components/ui/accordion';
-import cn from '@lib/utils';
-import { ReactNode } from 'react';
-import { useFormContext } from 'react-hook-form';
-import MenuItemContainer, { MenuItemContainerProps } from './MenuItemContainer';
-
-export type Date = {
-  endElement: ReactNode;
-  text: string;
-};
-
-export type AdvancedDateMobileFieldProps = {
-  dates: Date[];
-  onSelect?: (value: string | number) => void;
-};
-
-const AdvancedDateMobileField = ({
-  dates,
-  onSelect,
-}: AdvancedDateMobileFieldProps) => {
-  const { setValue } = useFormContext();
-
-  const handleSelect: MenuItemContainerProps['onSelect'] = (selected) => {
-    setValue('date', selected);
-    onSelect?.(selected);
-  };
-
-  return (
-    <Accordion type="single" collapsible>
-      <AccordionItem value="item-1">
-        <AccordionTrigger
-          className={cn(
-            ' bg-surface px-3 py-6',
-            'min-w-[21.9375rem] border-0 text-xl font-bold '
-          )}
-        >
-          日期
-        </AccordionTrigger>
-        <AccordionContent className="">
-          <MenuItemContainer items={dates} onSelect={handleSelect} />
-        </AccordionContent>
-      </AccordionItem>
-    </Accordion>
-  );
-};
-
-export default AdvancedDateMobileField;
