+++ src/components/SearchBar/fields/EventTypeFieldMenu.tsx
@@ -1,5 +1,11 @@
 'use client';
 
+import {
+  Accordion,
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
+} from '@components/ui/accordion';
 import cn from '@lib/utils';
 import {
   Popover,
@@ -9,12 +15,13 @@ import {
 import { motion } from 'framer-motion';
 import { ReactNode, useState } from 'react';
 import { useFormContext } from 'react-hook-form';
-import MenuItemContainer from './MenuItemContainer';
+import MenuItemContainer, { MenuItemContainerProps } from './MenuItemContainer';
 import menuAnimationVariants from './utils';
 
 export type Event = {
-  url: string;
+  url?: string;
   text: string;
+  endElement?: ReactNode;
 };
 
 export type EventTypeFieldMenuProps = {
@@ -89,4 +96,39 @@ const EventTypeFieldMenu = ({
     </Popover>
   );
 };
+
+export type AdvancedEventTypeMobileFieldProps = {
+  events: Event[];
+  onSelect?: (value: string | number) => void;
+};
+export const AdvancedEventTypeMobileField = ({
+  events,
+  onSelect,
+}: AdvancedEventTypeMobileFieldProps) => {
+  const { setValue } = useFormContext();
+
+  const handleSelect: MenuItemContainerProps['onSelect'] = (selected) => {
+    setValue('type', selected);
+    onSelect?.(selected);
+  };
+
+  return (
+    <Accordion type="single" collapsible>
+      <AccordionItem value="item-1">
+        <AccordionTrigger
+          className={cn(
+            ' bg-surface px-3 py-6',
+            'min-w-[21.9375rem] border-0 text-xl font-bold '
+          )}
+        >
+          形式
+        </AccordionTrigger>
+        <AccordionContent className="">
+          <MenuItemContainer items={events} onSelect={handleSelect} />
+        </AccordionContent>
+      </AccordionItem>
+    </Accordion>
+  );
+};
+
 export default EventTypeFieldMenu;
