+++ src/components/SearchBar/fields/utils.ts
@@ -11,23 +11,18 @@ import {
 
 export const locations = [
   {
-    url: '/',
     text: '北部',
   },
   {
-    url: '/',
     text: '中部',
   },
   {
-    url: '/',
     text: '南部',
   },
   {
-    url: '/',
     text: '東部',
   },
   {
-    url: '/',
     text: '離島',
   },
 ];
@@ -35,42 +30,34 @@ export const locations = [
 export const categories = [
   {
     icon: TreesIcon,
-    url: 'https://www.google.com/',
     text: '戶外踏青',
   },
   {
     icon: PartyPopperIcon,
-    url: '/',
     text: '社交活動',
   },
   {
     icon: HeartIcon,
-    url: '/',
     text: '興趣嗜好',
   },
   {
     icon: DumbbellIcon,
-    url: '/',
     text: '運動健身',
   },
   {
     icon: HazeIcon,
-    url: '/',
     text: '健康生活',
   },
   {
     icon: BotIcon,
-    url: '/',
     text: '科技玩物',
   },
   {
     icon: PaletteIcon,
-    url: '/',
     text: '藝術文化',
   },
   {
     icon: Gamepad2Icon,
-    url: '/',
     text: '遊戲',
   },
 ];
